[ed-application] cleaned unused dtos, functions and dependencies. Modify router address in diagnosis and feedback.
