--- diagnosis/src/main/java/swp/charite/diagnosis/dto/PatientDto.java
@@ -1,17 +1,0 @@
-package swp.charite.diagnosis.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class PatientDto {
-
-    private String firstname;
-    private String surname;
-    
-}
