+++ diagnosis/src/main/java/swp/charite/diagnosis/service/GuidanceService.java
@@ -13,7 +13,6 @@ import swp.charite.diagnosis.dto.GuidanceToPatient;
 import swp.charite.diagnosis.dto.GuidanceCreateEventDto;
 import swp.charite.diagnosis.model.Guidance;
 import swp.charite.diagnosis.model.OutboxEntity;
-import swp.charite.diagnosis.model.Priority;
 import swp.charite.diagnosis.repository.GuidanceRepository;
 import swp.charite.diagnosis.repository.OutboxRepository;
 
@@ -58,12 +57,4 @@ public class GuidanceService {
         }
     }
 
-    public void update(Long g_id) {
-        if (guidanceRepository.existsById(g_id)) {
-            Date date = new Date();
-            Guidance oldGuidance = guidanceRepository.findByGuidanceId(g_id);
-            oldGuidance.setDate(date.toString());
-            oldGuidance.setDone(true);
-        }
-    }
 }
