+++ feedbacks/src/main/java/swp/charite/feedbacks/controller/GuidanceController.java
@@ -7,17 +7,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import swp.charite.feedbacks.dto.FeedbackDto;
 import swp.charite.feedbacks.service.GuidanceCommandService;
 
 @RestController
-@RequestMapping(value = "/guidance")
+@RequestMapping(value = "/feedback")
 public class GuidanceController {
 
     @Autowired
     GuidanceCommandService guidanceCommandService;
 
-    @GetMapping("/done/{id}")
+    @GetMapping("/guidance/done/{id}")
     public ResponseEntity<?> markGuidanceAsDone(@PathVariable Long id) {
         if (guidanceCommandService.markGuidance(id)) {
             return new ResponseEntity<>(HttpStatus.OK);
