+++ feedbacks/src/main/java/swp/charite/feedbacks/dto/GetFeedbackDto.java
@@ -1,4 +1,5 @@
-package swp.charite.diagnosis.dto;
+package swp.charite.feedbacks.dto;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -6,11 +7,9 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@NoArgsConstructor
 @AllArgsConstructor
-public class DoctorDto {
-    
-    private String firstname;
-    private String surname;
-    
+@NoArgsConstructor
+public class GetFeedbackDto {
+    private String feedback;
+    private String date;
 }
