--- feedbacks/src/main/java/swp/charite/feedbacks/dto/_Feedback.java
@@ -1,15 +1,0 @@
-package swp.charite.feedbacks.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class _Feedback {
-    private String feedback;
-    private String date;
-}
